# improve GraphicsWidget paint speed by caching the bounding rect and the path used by the shape method