By caching the bounding rect the painting is increased significantly. On a live updating plot with ~160 curves the timings show the following:

#### With this patch 0.95 us/call
![image](https://user-images.githubusercontent.com/20952040/153834091-418ab5d9-1839-4073-9435-cb3d3b1c3ccc.png)

#### Without 1.99 us/call
![image](https://user-images.githubusercontent.com/20952040/153834894-568e5453-a2fc-4c0d-86de-848fb7c17d10.png)

Overall the painting speed went from 43.5ms/frame to 39.3ms/frame