@danielhrisca 2022-02-14
I've also added the caching of the path used by the ``shape`` method.

With ``shape`` caching:

``shape`` takes 0.18 us/call and the painting takes 36ms/frame
![image](https://user-images.githubusercontent.com/20952040/153848355-10f9ba39-3bd9-4fab-87a9-cf9ac8232ad3.png)

Without ``shape caching
``shape`` takes 2.48 us/call and the painting takes 39ms/frame
![image](https://user-images.githubusercontent.com/20952040/153848964-c667df2a-6de1-4439-9884-3af4429b6901.png)