@j9ac9k 2022-02-22
hi @danielhrisca 

Thanks for this PR and the profiling results.   I would ask you rename the variables to better reflect that they hold cached values.  This LGTM otherwise!